requirements.txt
ultralytics>=8.3.0
opencv-python>=4.5.0
pyside6>=6.0.0
numpy>=1.20.0
pillow>=9.0.0
torch>=2.0.0